+++ packages/xchain-thorchain-amm/src/utils.ts
@@ -1,9 +1,15 @@
-import { AssetAVAX } from '@xchainjs/xchain-avax'
-import { AssetBNB } from '@xchainjs/xchain-binance'
-import { AssetBSC } from '@xchainjs/xchain-bsc'
-import { AssetATOM } from '@xchainjs/xchain-cosmos'
-import { AssetETH } from '@xchainjs/xchain-ethereum'
-import { Asset, Chain, eqAsset } from '@xchainjs/xchain-util'
+import { AVAXChain, AssetAVAX, Client as AvaxClient, defaultAvaxParams } from '@xchainjs/xchain-avax'
+import { AssetBNB, BNBChain, Client as BnbClient } from '@xchainjs/xchain-binance'
+import { BTCChain, Client as BtcClient, defaultBTCParams as defaultBtcParams } from '@xchainjs/xchain-bitcoin'
+import { BCHChain, Client as BchClient, defaultBchParams } from '@xchainjs/xchain-bitcoincash'
+import { AssetBSC, BSCChain, Client as BscClient, defaultBscParams } from '@xchainjs/xchain-bsc'
+import { Network } from '@xchainjs/xchain-client'
+import { AssetATOM, Client as GaiaClient, GAIAChain } from '@xchainjs/xchain-cosmos'
+import { Client as DogeClient, DOGEChain, defaultDogeParams } from '@xchainjs/xchain-doge'
+import { AssetETH, Client as EthClient, ETHChain, defaultEthParams } from '@xchainjs/xchain-ethereum'
+import { Client as LtcClient, LTCChain, defaultLtcParams } from '@xchainjs/xchain-litecoin'
+import { Client as ThorClient, THORChain, defaultClientConfig as defaultThorParams } from '@xchainjs/xchain-thorchain'
+import { Address, Asset, Chain, eqAsset } from '@xchainjs/xchain-util'
 
 /**
  * Check if a chain is EVM and supported by the protocol
@@ -33,4 +39,31 @@ export const isProtocolERC20Asset = (asset: Asset): boolean => {
  */
 export const isProtocolBFTChain = (chain: Chain): boolean => {
   return [AssetBNB.chain, AssetATOM.chain].includes(chain)
+}
+
+export const validateAddress = (network: Network, chain: Chain, address: Address): boolean => {
+  switch (chain) {
+    case BTCChain:
+      return new BtcClient({ ...defaultBtcParams, network }).validateAddress(address)
+    case BCHChain:
+      return new BchClient({ ...defaultBchParams, network }).validateAddress(address)
+    case LTCChain:
+      return new LtcClient({ ...defaultLtcParams, network }).validateAddress(address)
+    case DOGEChain:
+      return new DogeClient({ ...defaultDogeParams, network }).validateAddress(address)
+    case ETHChain:
+      return new EthClient({ ...defaultEthParams, network }).validateAddress(address)
+    case AVAXChain:
+      return new AvaxClient({ ...defaultAvaxParams, network }).validateAddress(address)
+    case BSCChain:
+      return new BscClient({ ...defaultBscParams, network }).validateAddress(address)
+    case GAIAChain:
+      return new GaiaClient({ network }).validateAddress(address)
+    case BNBChain:
+      return new BnbClient({ network }).validateAddress(address)
+    case THORChain:
+      return new ThorClient({ ...defaultThorParams, network }).validateAddress(address)
+    default:
+      throw Error('Unsupported chain')
+  }
 }