+++ examples/liquidity/CHANGELOG.md
@@ -1,5 +1,18 @@
 # xchainjs-liquidity
 
+## 1.0.16
+
+### Patch Changes
+
+- Updated dependencies [99825cb]
+- Updated dependencies [6fe2b21]
+  - @xchainjs/xchain-thornode@0.3.16
+  - @xchainjs/xchain-util@0.13.7
+  - @xchainjs/xchain-thorchain-query@0.7.15
+  - @xchainjs/xchain-client@0.16.8
+  - @xchainjs/xchain-midgard-query@0.1.19
+  - @xchainjs/xchain-thorchain-amm@1.1.15
+
 ## 1.0.15
 
 ### Patch Changes