+++ packages/xchain-binance/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 5.7.17
+
+### Patch Changes
+
+- Updated dependencies [6fe2b21]
+  - @xchainjs/xchain-util@0.13.7
+  - @xchainjs/xchain-client@0.16.8
+
 ## 5.7.16
 
 ### Patch Changes